quick quick fix (#6584)

+++ testing/ui_tests/pages/base_page.py
@@ -2,7 +2,7 @@ import logging
 from ui_tests.pages.page import Page
 
 # will be 1 hour (Because of the interval)
-DISCOVERY_TIMEOUT = 60 * 60 * 5 * 0.4
+DISCOVERY_TIMEOUT = int(60 * 60 * 5 * 0.4)
 
 logger = logging.getLogger(f'axonius.{__name__}')
